+++ tests/unit/test_cli.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:49df91e6af9072a80c0fcd0d7c461f8316d671a68634189f96c4423f2a5a9539
-size 1335
+oid sha256:6f4130980eb8f2ae5f2b8ea73de7c0655ff7420adc6cb5dab0faf08fe95cca76
+size 1347