+++ tests/unit/test_io/test_packet_ingest.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2a18a87e38fc7b7d2901b3e27bd6b83105178b7dff0ee2c1614a481a997f3830
-size 8555
+oid sha256:7727c60f097983b304fa1f4316a52798b98cb52f370a56b8643fcefcfa8f15ba
+size 9122